Fixed some compiler warnings with latest gcc.
git-svn-id: https://svn.tdr.iem.uni-due.de:8443/svn/sctp/trunk/rsplib2@2430 516812f1-d3fd-0310-9ea4-9aededebcb87

+++ rsplib/environmentcache.cc
@@ -154,18 +154,18 @@ void EnvironmentCache::cleanUp()
       DIR* directory = opendir(CacheDirectory.c_str());
       if(directory != NULL) {
          char* cwd = getcwd(NULL, 0);
-         chdir(CacheDirectory.c_str());
+         if(chdir(CacheDirectory.c_str()) == 0) {
+            dirent* dentry = readdir(directory);
+            while(dentry != NULL) {
+               unlink(dentry->d_name);
+               dentry = readdir(directory);
+            }
+            closedir(directory);
 
-         dirent* dentry = readdir(directory);
-         while(dentry != NULL) {
-            unlink(dentry->d_name);
-            dentry = readdir(directory);
-         }
-         closedir(directory);
-
-         if(cwd) {
-            chdir(cwd);
-            free(cwd);
+            if(cwd) {
+               chdir(cwd);
+               free(cwd);
+            }
          }
       }
    }
